Fixed more MSVC compiler warnings

+++ Emulator/Misc/RetroShell/Interpreter.cpp
@@ -318,9 +318,9 @@ Interpreter::help(const Command& current)
     usage(current);
 
     // Determine tabular positions to align the output
-    isize tab = 0;
+    usize tab = 0;
     for (auto &it : current.subCommands) {
-        tab = std::max(tab, (isize)it.fullName.length());
+        tab = std::max(tab, it.fullName.length());
     }
     tab += indent.size();
